+++ tests/unit/test_solvers/test_idaklu_solver.py
@@ -13,14 +13,14 @@ from tests import get_discretisation_for_testing
 
 
 @pytest.mark.cibw
-@unittest.skipIf(not pybamm.have_idaklu(), "idaklu solver is not installed")
+@unittest.skipIf(not pybamm.has_idaklu(), "idaklu solver is not installed")
 class TestIDAKLUSolver(unittest.TestCase):
     def test_ida_roberts_klu(self):
         # this test implements a python version of the ida Roberts
         # example provided in sundials
         # see sundials ida examples pdf
         for form in ["casadi", "iree"]:
-            if (form == "iree") and (not pybamm.have_jax() or not pybamm.have_iree()):
+            if (form == "iree") and (not pybamm.has_jax() or not pybamm.has_iree()):
                 continue
             if form == "casadi":
                 root_method = "casadi"
@@ -66,7 +66,7 @@ class TestIDAKLUSolver(unittest.TestCase):
 
     def test_model_events(self):
         for form in ["casadi", "iree"]:
-            if (form == "iree") and (not pybamm.have_jax() or not pybamm.have_iree()):
+            if (form == "iree") and (not pybamm.has_jax() or not pybamm.has_iree()):
                 continue
             if form == "casadi":
                 root_method = "casadi"
@@ -187,7 +187,7 @@ class TestIDAKLUSolver(unittest.TestCase):
     def test_input_params(self):
         # test a mix of scalar and vector input params
         for form in ["casadi", "iree"]:
-            if (form == "iree") and (not pybamm.have_jax() or not pybamm.have_iree()):
+            if (form == "iree") and (not pybamm.has_jax() or not pybamm.has_iree()):
                 continue
             if form == "casadi":
                 root_method = "casadi"
@@ -245,9 +245,7 @@ class TestIDAKLUSolver(unittest.TestCase):
     def test_sensitivities_initial_condition(self):
         for form in ["casadi", "iree"]:
             for output_variables in [[], ["2v"]]:
-                if (form == "iree") and (
-                    not pybamm.have_jax() or not pybamm.have_iree()
-                ):
+                if (form == "iree") and (not pybamm.has_jax() or not pybamm.has_iree()):
                     continue
                 if form == "casadi":
                     root_method = "casadi"
@@ -298,7 +296,7 @@ class TestIDAKLUSolver(unittest.TestCase):
         # example provided in sundials
         # see sundials ida examples pdf
         for form in ["casadi", "iree"]:
-            if (form == "iree") and (not pybamm.have_jax() or not pybamm.have_iree()):
+            if (form == "iree") and (not pybamm.has_jax() or not pybamm.has_iree()):
                 continue
             if form == "casadi":
                 root_method = "casadi"
@@ -404,7 +402,7 @@ class TestIDAKLUSolver(unittest.TestCase):
         # example provided in sundials
         # see sundials ida examples pdf
         for form in ["casadi", "iree"]:
-            if (form == "iree") and (not pybamm.have_jax() or not pybamm.have_iree()):
+            if (form == "iree") and (not pybamm.has_jax() or not pybamm.has_iree()):
                 continue
             if form == "casadi":
                 root_method = "casadi"
@@ -446,7 +444,7 @@ class TestIDAKLUSolver(unittest.TestCase):
         # example provided in sundials
         # see sundials ida examples pdf
         for form in ["casadi", "iree"]:
-            if (form == "iree") and (not pybamm.have_jax() or not pybamm.have_iree()):
+            if (form == "iree") and (not pybamm.has_jax() or not pybamm.has_iree()):
                 continue
             if form == "casadi":
                 root_method = "casadi"
@@ -605,7 +603,7 @@ class TestIDAKLUSolver(unittest.TestCase):
 
     def test_dae_solver_algebraic_model(self):
         for form in ["casadi", "iree"]:
-            if (form == "iree") and (not pybamm.have_jax() or not pybamm.have_iree()):
+            if (form == "iree") and (not pybamm.has_jax() or not pybamm.has_iree()):
                 continue
             if form == "casadi":
                 root_method = "casadi"
@@ -914,7 +912,7 @@ class TestIDAKLUSolver(unittest.TestCase):
         # equivalence
 
         for form in ["casadi", "iree"]:
-            if (form == "iree") and (not pybamm.have_jax() or not pybamm.have_iree()):
+            if (form == "iree") and (not pybamm.has_jax() or not pybamm.has_iree()):
                 continue
             if form == "casadi":
                 root_method = "casadi"
@@ -1087,7 +1085,7 @@ class TestIDAKLUSolver(unittest.TestCase):
 
     def test_python_idaklu_deprecation_errors(self):
         for form in ["python", "", "jax"]:
-            if form == "jax" and not pybamm.have_jax():
+            if form == "jax" and not pybamm.has_jax():
                 continue
 
             model = pybamm.BaseModel()